requirements.txt
ConfigArgParse==1.2.3
dataclasses==0.8
fasttext==0.9.2
Flask-PyMongo==2.3.0
google-cloud-texttospeech==1.0.1
librosa==0.8.0
lmdb==1.0.0
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
pyarrow==0.15.0
pymongo==3.11.0
SoundFile==0.10.3.post1
scikit-learn==0.23.2
scipy==1.5.4
tabulate==0.8.7
torch==1.7.1
torchvision==0.8.2
tqdm==4.51.0
umap==0.1.1
waitress==1.4.4
wheel==0.35.1